refactor(react): split library file to vcn and Slot
